Второе ревью ревью 12 тз

+++ src/main/java/ru/yandex/practicum/filmorate/service/UserService.java
@@ -5,11 +5,13 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
+import ru.yandex.practicum.filmorate.exception.NotFoundException;
 import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.User;
 import ru.yandex.practicum.filmorate.storage.user.UserStorage;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @Slf4j
@@ -19,7 +21,11 @@ public class UserService {
 
     public List<User> getUsers() {
         log.info("Запрос на получение всех пользователей");
-        return userStorage.getUsers();
+        List<User> users = userStorage.getUsers();
+        for (User user : users) {
+            user.setFriends(userStorage.getUserFriendsId(user.getId()));
+        }
+        return users;
     }
 
     public User createUser(User user) {
@@ -41,33 +47,47 @@ public class UserService {
             newUser.setName(newUser.getLogin());
             log.info("Обновление. Пустое имя пользователя, использован Логин {}", newUser.getLogin());
         }
-        userStorage.getUserById(newUser.getId());
+        if (!userStorage.checkIfUserExists(newUser.getId())) {
+            throw new NotFoundException("Пользователь с id " + newUser.getId() + " не найден");
+        }
         return userStorage.updateUser(newUser);
     }
 
     public User getUserById(Long id) {
         log.info("Запрос на получение пользователя с Id {}", id);
-        return userStorage.getUserById(id);
+        Optional<User> userOptional = userStorage.getUserById(id);
+        User user = userOptional.orElseThrow(() -> new NotFoundException("пользователь с id " + id + " не найден"));
+        user.setFriends(userStorage.getUserFriendsId(user.getId()));
+        return user;
     }
 
     public void addFriend(Long id, Long friendId) {
-        userStorage.getUserById(id);
-        userStorage.getUserById(friendId);
+        if (!userStorage.checkIfUserExists(id)) {
+            throw new NotFoundException("Пользователь с id " + id + " не найден");
+        }
+        if (!userStorage.checkIfUserExists(friendId)) {
+            throw new NotFoundException("Пользователь с id " + friendId + " не найден");
+        }
         userStorage.addFriend(id, friendId);
         log.info("Пользователь с Id {} добавил в друзья пользователя с Id {}", id, friendId);
     }
 
     public void deleteFriend(Long userId, Long friendId) {
-        userStorage.getUserById(userId);
-        userStorage.getUserById(friendId);
-
+        if (!userStorage.checkIfUserExists(userId)) {
+            throw new NotFoundException("Пользователь с id " + userId + " не найден");
+        }
+        if (!userStorage.checkIfUserExists(friendId)) {
+            throw new NotFoundException("Пользователь с id " + friendId + " не найден");
+        }
         userStorage.deleteFriend(userId, friendId);
         log.info("Пользователь с Id {} удалил из друзей пользователя с Id {}", userId, friendId);
     }
 
     public List<User> getUserFriends(Long id) {
         log.info("Запрос на получение всех друзей пользователя с id {}", id);
-        userStorage.getUserById(id);
+        if (!userStorage.checkIfUserExists(id)) {
+            throw new NotFoundException("Пользователь с id " + id + " не найден");
+        }
         return userStorage.getUserFriends(id);
     }
